move delete order route from main route
