+++ app/routes.js
@@ -11,10 +11,9 @@ import { getTaskListPageContext } from './pages/task-list/controller';
 import { sectionRoutes } from './pages/sections/routes';
 import { summaryRoutes } from './pages/summary/routes';
 import { completeOrderRoutes } from './pages/complete-order/routes';
+import { deleteOrderRoutes } from './pages/delete-order/routes';
 import includesContext from './includes/manifest.json';
 import { clearSession } from './helpers/routes/sessionHelper';
-import { getDeleteOrderContext, deleteOrder } from './pages/delete-order/controller';
-import { getDeleteOrderConfirmationContext } from './pages/delete-order/delete-order-confirmation/controller';
 
 const addContext = ({ context, user, csrfToken }) => ({
   ...context,
@@ -72,44 +71,7 @@ export const routes = (authProvider, sessionManager) => {
 
   router.use('/organisation/:orderId/complete-order', completeOrderRoutes(authProvider, addContext, sessionManager));
 
-  router.get('/organisation/:orderId/delete-order', authProvider.authorise({ claim: 'ordering' }), withCatch(logger, authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const accessToken = extractAccessToken({ req, tokenType: 'access' });
-
-    const context = await getDeleteOrderContext({
-      req,
-      sessionManager,
-      accessToken,
-      logger,
-    });
-
-    logger.info(`navigating to order ${orderId} delete-order page`);
-    res.render('pages/delete-order/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.post('/organisation/:orderId/delete-order', authProvider.authorise({ claim: 'ordering' }), withCatch(logger, authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const accessToken = extractAccessToken({ req, tokenType: 'access' });
-
-    await deleteOrder({ orderId, accessToken });
-
-    return res.redirect(`${config.baseUrl}/organisation/${orderId}/delete-order/confirmation`);
-  }));
-
-  router.get('/organisation/:orderId/delete-order/confirmation', authProvider.authorise({ claim: 'ordering' }), withCatch(logger, authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const accessToken = extractAccessToken({ req, tokenType: 'access' });
-
-    const context = await getDeleteOrderConfirmationContext({
-      req,
-      sessionManager,
-      accessToken,
-      logger,
-    });
-
-    logger.info(`navigating to order ${orderId} delete-order-confirmation page`);
-    res.render('pages/delete-order/delete-order-confirmation/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
+  router.use('/organisation/:orderId/delete-order', deleteOrderRoutes(authProvider, addContext, sessionManager));
 
   router.use('/organisation/:orderId', sectionRoutes(authProvider, addContext, sessionManager));
